create graph_comp.py and move graph analysis functions from graph_gen.py and strat_comp.py there
