+++ src/graph_gen.py
@@ -1,188 +1,7 @@
-# Functionality related to generating, encoding, and decoding binary adjacency matrices 
-# for a variety of environment graphs 
+# Functionality related to generating binary adjacency matrices for a variety of environment graphs 
 import jax
 import jax.numpy as jnp
 import numpy as np
-
-def gen_graph_code(A):
-    """
-    Generate a unique code representing the environment graph.
-
-    Parameters
-    ----------
-    A : jaxlib.xla_extension.DeviceArray 
-        Binary adjacency matrix of the environment graph.
-
-    Returns
-    -------
-    String
-        Unique encoding of the binary adjacency matrix.
-
-    """
-    bin_string = "1"  # leading 1 added to avoid issues with decimal conversion in decoding
-    for i in range(jnp.shape(A)[0] - 1):
-        for j in range(i + 1, jnp.shape(A)[1]):
-            bin_string = bin_string + str(int(A[i, j]))
-    graph_num = int(bin_string, base=2)
-    graph_code = "N" + str(A.shape[0]) + "_" + str(graph_num)
-    return graph_code
-
-def gen_graph_hexcode(A):
-    """
-    Generate a unique hexadecimal code representing the environment graph.
-
-    Parameters
-    ----------
-    A : numpy.ndarray 
-        Binary adjacency matrix of the environment graph.
-
-    Returns
-    -------
-    String
-        Unique hexadecimal encoding of the binary adjacency matrix.
-
-    """
-    bin_string = "1" # leading 1 added to avoid issues with decimal conversion in decoding
-    for i in range(A.shape[0] - 1):
-        for j in range(i + 1, A.shape[1]):
-            bin_string = bin_string + str(int(A[i, j]))
-    graph_num = int(bin_string, base=2)
-    graph_hex = hex(graph_num)
-    graph_code = "N" + str(A.shape[0]) + "_" + str(graph_hex).lstrip("0x")
-    return graph_code
-
-def graph_decode(graph_code):
-    """
-    Generate binary adjacency matrix of environment graph from graph code.
-
-    Parameters
-    ----------
-    graph_code : String
-        Unique encoding of the binary adjacency matrix. 
-
-    Returns
-    -------
-    jaxlib.xla_extension.DeviceArray
-        Binary adjacency matrix of the environment graph.
-    
-    """
-    n = int(graph_code[1:graph_code.find("_")])
-    U_dec = int(graph_code[(graph_code.find("_") + 1):])
-    U_bin = bin(U_dec)
-    U_bin_list = list(U_bin[(U_bin.find("b") + 2):]) # skip leading 1 in binary string
-    U_bin_list = [int(x) for x in U_bin_list] 
-    inds = jnp.triu_indices(n, 1)
-    U = jnp.zeros((n, n))
-    U = U.at[inds].set(U_bin_list)
-    A = U + jnp.transpose(U) + jnp.identity(n)
-    return A
-
-def graph_diam(A):
-    """
-    Compute diameter of graph described by binary adjacency matrix `A`.
-
-    Parameters
-    ----------
-    A : jaxlib.xla_extension.DeviceArray
-        Binary adjacency matrix of the environment graph.
-
-    Returns
-    -------
-    int
-        Diameter of the environment graph.
-    
-    """
-    n = jnp.shape(A)[0]
-    if jnp.all(A != 0):
-        return 1
-    A_ = A
-    for i in range(2, n):
-        A_ = jnp.matmul(A, A_)
-        if jnp.all(A_ != 0):
-            return i
-    return np.NaN
-
-def get_leaf_node_pairs(A):
-    """
-    Get the set of all leaf node pairs for the given environment graph.
-
-    Parameters
-    ----------
-    A : jaxlib.xla_extension.DeviceArray
-        Binary adjacency matrix of the environment graph.
-
-    Returns
-    -------
-    jaxlib.xla_extension.DeviceArray
-        (D x 2) array with each row containing an ordered leaf node pair,
-        where D is the total number of ordered leaf pairs within the graph. 
-    
-    """
-    leaf_inds = jnp.sum(A, axis=1) == 2
-    leaf_map = jnp.outer(leaf_inds, leaf_inds)
-    leaf_node_pairs = jnp.argwhere(leaf_map)
-    return leaf_node_pairs
-
-def get_diametric_pairs(A):
-    """
-    Get the set of all leaf node pairs separated by the graph diameter
-    for environment graph given by binary adjacency matrix `A`.
-
-    Parameters
-    ----------
-    A : jaxlib.xla_extension.DeviceArray
-        Binary adjacency matrix of the environment graph.
-
-    Returns
-    -------
-    jaxlib.xla_extension.DeviceArray
-        (D x 2) array with each row containing an ordered diametric leaf node pair,
-        where D is the total number of ordered diametric pairs within the graph. 
-    
-    """
-    dg = graph_diam(A)
-    diam_pairs = []
-    A_ = A
-    for k in range(2, dg):
-        A_ = jnp.matmul(A, A_)
-    diam_pairs = jnp.argwhere(A_ == 0)
-    return diam_pairs
-
-def get_shortest_path_distances(A, node_pairs):
-    """
-    Compute the shortest-path distance between given node pairs
-    for environment graph given by binary adjacency matrix `A`.
-
-    Considers distance from a node to itself to be 1, for consistency 
-    with the robotic surveillance problem formulation. 
-    For details, see https://arxiv.org/abs/2011.07604.
-
-    Parameters
-    ----------
-    A : jaxlib.xla_extension.DeviceArray
-        Binary adjacency matrix of the environment graph.
-    node_pairs : jaxlib.xla_extension.DeviceArray  
-        (D x 2) array of pairs of nodes. 
-
-
-    Returns
-    -------
-    jaxlib.xla_extension.DeviceArray
-        (D x 3) array with each row containing a node pair, with corresponding
-        shortest-path distance in the third column. 
-    """
-    node_pair_SPDs = jnp.zeros((jnp.shape(node_pairs)[0], 3), dtype=int)
-    node_pair_SPDs = node_pair_SPDs.at[:, :2].set(node_pairs)
-    n = jnp.shape(A)[0]
-    A_ = A
-    for k in range(1, n):
-        k_spd_pairs = jnp.argwhere(A_ != 0)
-        for pair_ind in range(jnp.shape(node_pair_SPDs)[0]):
-            if node_pair_SPDs[pair_ind, 2] == 0:
-                if jnp.shape(jnp.argwhere((k_spd_pairs[:, 0] == node_pair_SPDs[pair_ind, 0]) * (k_spd_pairs[:, 1] == node_pair_SPDs[pair_ind, 1])))[0] != 0:
-                    node_pair_SPDs = node_pair_SPDs.at[pair_ind, 2].set(k)
-        A_ = jnp.matmul(A, A_)
-    return node_pair_SPDs
 
 def gen_star_G(n):
     """
