+++ src/strat_comp.py
@@ -544,82 +544,6 @@ def comp_avg_LCP_grad_param(Q, A, F0, tau, num_LCPs=None):
     grad = jnp.mean(J, axis=0)
     return grad
 
-
-def get_closest_sym_strat_grid(P_ref, P_comp, gridrows, gridcols, sym_index=None):
-    if(gridrows == gridcols):
-        P_syms = jnp.stack([P_comp, sq_grid_rot90(P_comp, gridrows, gridcols), grid_rot180(P_comp), \
-                sq_grid_rot270(P_comp, gridrows, gridcols), grid_row_reflect(P_comp, gridrows, gridcols), \
-                grid_col_reflect(P_comp, gridrows, gridcols), sq_grid_transpose(P_comp, gridrows, gridcols), \
-                sq_grid_antitranspose(P_comp, gridrows, gridcols)
-                ])
-        sum_sq_diffs = jnp.full(8, np.Inf)
-    elif(gridrows == 1 or gridcols == 1):
-        P_syms = jnp.stack([P_comp, grid_rot180(P_comp)])
-        sum_sq_diffs = jnp.full(2, np.Inf)
-    else:
-        P_syms = jnp.stack([P_comp, grid_rot180(P_comp), grid_row_reflect(P_comp, gridrows, gridcols), \
-                                grid_col_reflect(P_comp, gridrows, gridcols)])
-        sum_sq_diffs = jnp.full(4, np.Inf)
-
-    if sym_index is not None:
-        P_closest = jnp.reshape(P_syms[sym_index, :, :], (gridrows*gridcols, gridrows*gridcols))
-    else:
-        for k in range(jnp.shape(P_syms)[0]):
-            sum_sq_diffs = sum_sq_diffs.at[k].set(jnp.sum((P_syms[k, :, :] - P_ref)**2))
-        sym_index = jnp.argwhere(sum_sq_diffs == jnp.min(sum_sq_diffs))
-        P_closest = jnp.reshape(P_syms[sym_index, :, :], (gridrows*gridcols, gridrows*gridcols))
-    return P_closest, sym_index
-
-def grid_row_reflect(M, gridrows, gridcols):
-    grid_perm = jnp.flipud(jnp.identity(gridrows)) # antidiagonal permutation matrix
-    block = jnp.identity(gridcols)
-    M_perm = jnp.kron(grid_perm, block)
-    M = jnp.matmul(M_perm, M)
-    M = jnp.matmul(M, jnp.transpose(M_perm))
-    return M
-
-def grid_col_reflect(M, gridrows, gridcols):
-    grid_perm = jnp.flipud(jnp.identity(gridrows)) # antidiagonal permutation matrix
-    block = jnp.identity(gridcols)
-    M_perm = jnp.kron(block, grid_perm)
-    M = jnp.matmul(jnp.transpose(M_perm), M)
-    M = jnp.matmul(M, M_perm)
-    return M
-
-def grid_rot180(M):  
-    M = jnp.fliplr(M)
-    M = jnp.flipud(M)
-    return M
-
-def sq_grid_transpose(M, gridrows, gridcols):
-    n = gridrows*gridcols
-    M_reflect = jnp.full((n, n), np.NaN)
-    identity_map = jnp.stack([jnp.outer(jnp.arange(n, dtype=int), jnp.ones(n, dtype=int)), \
-                                jnp.outer(jnp.ones(n, dtype=int), jnp.arange(n, dtype=int))])
-    node_map = jnp.flipud(jnp.arange(0, n).reshape(gridrows, gridcols))
-    ref_map = jnp.flipud(jnp.transpose(node_map)).flatten()
-    reflect_map = jnp.stack([jnp.outer(ref_map, jnp.ones(n, dtype=int)), \
-                                jnp.outer(jnp.ones(n, dtype=int), ref_map)])
-    M_reflect = M_reflect.at[identity_map[0, :, :], identity_map[1, :, :]].set(M[reflect_map[0, :, :], reflect_map[1, :, :]])
-    return M_reflect
-
-def sq_grid_antitranspose(M, gridrows, gridcols):
-    M = sq_grid_transpose(M, gridrows, gridcols)
-    M = grid_rot180(M)
-    return M
-
-def sq_grid_rot90(M, gridrows, gridcols):
-    M = sq_grid_transpose(M, gridrows, gridcols)
-    M = grid_row_reflect(M, gridrows, gridcols)
-    return M
-
-def sq_grid_rot270(M, gridrows, gridcols):
-    M = grid_row_reflect(M, gridrows, gridcols)
-    M = sq_grid_transpose(M, gridrows, gridcols)
-    return M
-
-
-
 # TESTING -----------------------------------------------------------------------------------------
 if __name__ == '__main__':
     np.set_printoptions(linewidth=np.inf)
